Make sure we climb even for `esy @name` invocations, add tests
